fix(traces-v2): move sidebar expand button to lens tab strip, lower collapse threshold
The footer expand button got buried under pagination chrome and missed
on a quick scan — the lens tab strip ("All / Conversations / Errors /
...") is where the operator's eye already lives. Anchor the expand
button to the LEFT of "All" instead, only when the sidebar is
collapsed. Drop the collapse threshold from 80px to 40px so a short
accidental drag doesn't snap the sidebar shut.
